Split core.classes into separate modules core.material, core.ray and core.intersection.
