+++ raysect/core/__init__.py
@@ -1,4 +1,6 @@
-from .classes import *
+from .ray import *
+from .intersection import *
+from .material import *
 from .boundingbox import *
 from .math import *
 from .scenegraph import *
