+++ raysect/core/acceleration/boundprimitive.pxd
@@ -32,8 +32,8 @@
 from raysect.core.boundingbox cimport BoundingBox3D
 from raysect.core.scenegraph.primitive cimport Primitive
 from raysect.core.math.point cimport Point3D
-from raysect.core.classes cimport Ray
-from raysect.core.classes cimport Intersection
+from raysect.core.ray cimport Ray
+from raysect.core.intersection cimport Intersection
 
 cdef class BoundPrimitive:
