+++ raysect/core/acceleration/kdtree.pxd
@@ -31,7 +31,7 @@
 
 from raysect.core.acceleration.accelerator cimport Accelerator as _Accelerator
 from raysect.core.math.spatial.kdtree3d cimport KDTree3DCore as _KDTreeCore
-from raysect.core.classes cimport Intersection
+from raysect.core.intersection cimport Intersection
 
 cdef class _PrimitiveKDTree(_KDTreeCore):
     cdef:
