+++ raysect/core/acceleration/kdtree.pyx
@@ -32,7 +32,7 @@
 from raysect.core.math.point cimport Point3D
 from raysect.core.math.spatial.kdtree3d cimport Item3D
 from raysect.core.scenegraph.primitive cimport Primitive
-from raysect.core.classes cimport Ray
+from raysect.core.ray cimport Ray
 from raysect.core.acceleration.boundprimitive cimport BoundPrimitive
 from libc.stdint cimport int32_t
 cimport cython
