+++ raysect/core/acceleration/unaccelerated.pyx
@@ -31,10 +31,10 @@
 
 # TODO: add docstrings
 cimport cython
-from raysect.core.classes cimport Ray
+from raysect.core.ray cimport Ray
 from raysect.core.scenegraph.primitive cimport Primitive
 from raysect.core.math.point cimport Point3D
-from raysect.core.classes cimport Intersection
+from raysect.core.intersection cimport Intersection
 from raysect.core.acceleration.boundprimitive cimport BoundPrimitive
 
 cdef class Unaccelerated(Accelerator):
