+++ raysect/core/material.pyx
@@ -29,52 +29,23 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from raysect.core.ray cimport Ray
-from raysect.core.intersection cimport Intersection
-from raysect.core.scenegraph.node cimport Node
-from raysect.core.scenegraph.primitive cimport Primitive
-from raysect.core.acceleration.boundprimitive cimport BoundPrimitive
+cdef class Material:
 
-cdef class CSGPrimitive(Primitive):
+    def __init__(self):
+        self.primitives = []
 
-    cdef CSGRoot _csgroot
-    cdef BoundPrimitive _primitive_a
-    cdef BoundPrimitive _primitive_b
-    cdef Ray _cache_ray
-    cdef Intersection _cache_intersection_a
-    cdef Intersection _cache_intersection_b
-    cdef Intersection _cache_last_intersection
-    cdef bint _cache_invalid
+    cpdef object notify_material_change(self):
+        """
+        Notifies any connected scene-graph root of a change to the material.
 
-    cdef inline Intersection _identify_intersection(self, Ray ray, Intersection intersection_a, Intersection intersection_b, Intersection closest_intersection)
+        The notification informs the root node that any caching structures used
+        to accelerate ray-tracing calculations are now potentially invalid and
+        must be recalculated, taking the new material properties into account.
+        """
 
-    cdef inline Intersection _closest_intersection(self, Intersection a, Intersection b)
+        for primitive in self.primitives:
+            primitive.notify_material_change()
 
-    cdef bint _valid_intersection(self, Intersection a, Intersection b, Intersection closest)
-
-    cdef Intersection _modify_intersection(self, Intersection closest, Intersection a, Intersection b)
-
-    cdef void rebuild(self)
-
-
-cdef class CSGRoot(Node):
-
-    cdef CSGPrimitive csg_primitive
-
-
-cdef class Union(CSGPrimitive):
-
-    pass
-
-
-cdef class Intersect(CSGPrimitive):
-
-    pass
-
-
-cdef class Subtract(CSGPrimitive):
-
-    pass
