+++ raysect/core/scenegraph/primitive.pxd
@@ -29,7 +29,9 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from raysect.core.classes cimport Ray, Material, Intersection
+from raysect.core.ray cimport Ray
+from raysect.core.material cimport Material
+from raysect.core.intersection cimport Intersection
 from raysect.core.math.point cimport Point3D
 from raysect.core.scenegraph.node cimport Node
 from raysect.core.boundingbox cimport BoundingBox3D
