+++ raysect/core/scenegraph/tests/test_primitive.py
@@ -31,7 +31,7 @@ import unittest
 from raysect.core.scenegraph import Primitive, Node
 from raysect.core.math import Point3D, AffineMatrix3D, translate
 from raysect.core import BoundingBox3D
-from raysect.core.classes import Material, Ray
+from raysect.core.ray import Material, Ray
 
 # TODO: Port to Cython to allow testing of the Cython API and allow access to internal structures
 # TODO: Add tests for functionality inherited from Node.
