+++ raysect/primitive/lens/spherical.pyx
@@ -33,7 +33,7 @@ from raysect.primitive import Intersect, Subtract, Sphere, Cylinder
 
 from raysect.core.math.affinematrix cimport AffineMatrix3D
 from raysect.core.math.transform cimport translate
-from raysect.core.classes cimport Material
+from raysect.core.material cimport Material
 from raysect.primitive.utility cimport EncapsulatedPrimitive
 from libc.math cimport sqrt
