+++ raysect/primitive/sphere.pyx
@@ -29,7 +29,8 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from raysect.core.classes cimport Material, new_intersection
+from raysect.core.material cimport Material
+from raysect.core.intersection cimport new_intersection
 from raysect.core.boundingbox cimport BoundingBox3D
 from raysect.core.math.point cimport new_point3d
 from raysect.core.math.normal cimport new_normal3d, Normal3D
